Fix bugs and set up #about section
When the site gets refreshed, the navigation bar somehow gets displayed
behind all the other elements for a few seconds.
Set z-index to highest in order to avoid this wrong behaviour.

Update button animation of .hire, so it is cleaner when hovering over.

Also set up second section #about, but still not finished.