Fix navbar animation when scrolling and Home Button of navbar is working again
Home button is not working. Some typo in section>id declaration.
So adjust this, then the button should be working fine.

Navbar animation when scrolling is not working, because the function window.onscroll is getting called twice in the code.
And for some reason, only the second function call is working, where the scroll-to-top button is operated.
So lets combine both function calls in one, so both functionalities can work properly again.